Markov Sampling and Simulation Pipeline

Markov Sampling and Simulation Pipeline

markov_simulation_pipeline(
  markov_model = NULL,
  duration = NULL,
  discount = 1.035,
  samples = 1,
  sample_type = "base",
  sim_type = "base",
  map_fn = NULL,
  debug = FALSE
)

Arguments

markov_model
A list of functions that define a markov model across multiple interventions. See example_two_state_markov for the correct format.

duration
Numeric, how many long to run the model for.

discount
Numeric, the discount that should be applied to the costs and QALYs. Defaults to 1.035.

samples
Numeric, defaults to 1. The number of markov model samples to use.

sample_type
A character string specifying the approach to use to sample the model. Options and defaults inherited from sample_markov.

sim_type
A character string specifying the approach to use to simulate the model. Options and defaults inherited from simulate_markov.

map_fn
An R function used to iterate over the model samples and run simulations. Must accept a functon as an argument. Defaults to using purrr::map if not supplied.

debug
Logical, defaults to FALSE. Turns on all debug checks - this may impact runtimes.

Value

A list containing the model samples and simulations.

See also

Examples

markov_simulation_pipeline(example_two_state_markov(), duration = 10, samples = 2)

#> # A tibble: 4 x 9
#>   sample intervention transition state_cost intervention_co… cohort qalys
#>    <int> <chr>        <named li> <named li>            <dbl> <name> <nam>
#> 1      1 SoC          <dbl[,2] … <dbl [2]>                 0 <dbl … <dbl…
#> 2      1 Soc with We… <dbl[,2] … <dbl [2]>                50 <dbl … <dbl…
#> 3      2 SoC          <dbl[,2] … <dbl [2]>                 0 <dbl … <dbl…
#> 4      2 Soc with We… <dbl[,2] … <dbl [2]>                50 <dbl … <dbl…
#> # … with 2 more variables: total_costs <dbl>, total_qalys <dbl>