Arrange Vectorised Vector Samples

A convenience function used to arrange vectorised samples into the correct vector format for several functions used to specify Markov model. See example_two_state_markov for an example use case.

vector_arrange(samples)

Arguments

samples
A list of vectorised samples

Value

A list of vectors with each vector representing a single sample.

Examples

vector_samples <- list(rnorm(10, 1, 2), rnorm(10, 10, 2))

vector_arrange(vector_samples)

#> [[1]]
#> [1] 2.570639 9.126159
#> 
#> [[2]]
#> [1] 1.416211 6.877566
#> 
#> [[3]]
#> [1] 1.519532 9.723292
#> 
#> [[4]]
#> [1] 1.368244 8.611735
#> 
#> [[5]]
#> [1] 0.9399837 9.1193757
#> 
#> [[6]]
#> [1] 3.997828 8.603714
#> 
#> [[7]]
#> [1]  1.66773 10.09906
#> 
#> [[8]]
#> [1] 1.481255 6.568329
#> 
#> [[9]]
#> [1]  1.663739 10.931949
#> 
#> [[10]]
#> [1]  2.151485 12.316740
#>